# Docs/Architettura/openapi.yaml
openapi: 3.0.0
info:
  title: RiciclApp
  description: Documentazione delle API per il progetto RiciclApp.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api  
    description: Server Locale (Sviluppo)
paths:
...
  /reports/create:
    post:
      summary: Crea un nuovo report
      description: Crea una segnalazione associata a un bidone e imposta lo stato del bidone a "SEGNALATO".
      tags:
        - Reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - binId
                - description
              properties:
                userId:
                  type: string
                  description: ID dell'utente che crea la segnalazione
                  example: "683a7e4f9b12c8a12345678"
                binId:
                  type: string
                  description: ID del bidone segnalato
                  example: "683a7e4f9b12c8a87654321"
                description:
                  type: string
                  description: Descrizione del problema
                  example: "Il bidone è danneggiato e non si chiude."
      responses:
        '201':
          description: Report creato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Report creato con successo
                  report:
                    $ref: '#/components/schemas/Report'

        '400':
          description: Descrizione mancante o non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La descrizione è obbligatoria

        '404':
          description: Bidone non trovato
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bidone non trovato

        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Errore interno del server durante la creazione del report
  /reports/all:
      get:
        summary: Recupera tutte le segnalazioni
        description: Restituisce l'elenco completo delle segnalazioni con informazioni sull'utente, sul bidone e sul centro di raccolta associato.
        tags:
          - Reports

        responses:
          '200':
            description: Lista delle segnalazioni recuperata con successo
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                      example: true

                    reports:
                      type: array

          '500':
            description: Errore interno del server
  /reports/update/{reportId}:
      put:
        summary: Aggiorna lo stato di una segnalazione
        description: |
          Aggiorna lo stato di una segnalazione.

          Comportamento:
          - ACCEPT → il bidone passa in MANUTENZIONE
          - RESOLVED → il bidone torna in OK e il report viene eliminato
          - REJECTED → il bidone torna in OK e il report viene eliminato
        tags:
          - Reports

        parameters:
          - in: path
            name: reportId
            required: true
            schema:
              type: string
            description: ID del report da aggiornare

        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    enum:
                      - ACCEPT
                      - RESOLVED
                      - REJECTED
                    example: ACCEPT

        responses:
          '200':
            description: Stato aggiornato con successo
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Stato del report aggiornato con successo

                    report:
                      $ref: '#/components/schemas/Report'

          '404':
            description: Report non trovato

          '500':
            description: Errore interno del server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      required:
        - userId
        - binId
        - description
        - status
      properties:
        _id:
          type: string
          description: ID univoco del report
          example: "683a7e4f9b12c8a99999999"

        userId:
          type: string
          description: ID dell'utente che ha creato la segnalazione
          example: "683a7e4f9b12c8a12345678"

        binId:
          type: string
          description: ID del bidone segnalato
          example: "683a7e4f9b12c8a87654321"

        description:
          type: string
          description: Descrizione del problema riscontrato
          example: "Il bidone è danneggiato e non si chiude."

        status:
          type: string
          description: Stato della segnalazione
          enum:
            - PENDING
            - IN_PROGRESS
            - RESOLVED
            - REJECTED
            - ASSIGNED
          example: PENDING

        assignedTo:
          type: string
          nullable: true
          description: Operatore assegnato alla gestione della segnalazione
          example: "683a7e4f9b12c8a55555555"

        createdAt:
          type: string
          format: date-time
          description: Data di creazione del report
          example: "2025-06-01T12:34:56.000Z"

        updatedAt:
          type: string
          format: date-time
          description: Data dell'ultima modifica del report
          example: "2025-06-01T12:34:56.000Z"